Split EventsTests into smaller files.
