+++ GitHubHook.Tests/Events/PingEventTests.cs
@@ -1,0 +1,25 @@
+using System;
+using GitHubHook.Events;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Newtonsoft.Json;
+
+namespace GitHubHook.Tests.Events
+{
+    [TestClass]
+    public class PingEventTests : BaseEventTests
+    {
+        [TestMethod]
+        public void Deserialize_ShouldSucceed()
+        {
+            // Arrange
+            var payload = resourceManager.GetString("PingPayload.json");
+
+            // Act
+            var pingPayload = JsonConvert.DeserializeObject<PingEvent>(payload);
+
+            // Assert
+            Assert.IsNotNull(pingPayload);
+            Console.WriteLine(JsonConvert.SerializeObject(pingPayload));
+        }
+    }
+}
